requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
torch>=1.9.0
gensim>=4.0.1
psycopg2-binary>=2.9.1
sqlalchemy>=1.4.0
transformers>=4.5.0
matplotlib>=3.4.0
seaborn>=0.11.0
jupyter>=1.0.0
tqdm>=4.61.0
wikipedia-api>=0.5.4 